adjust stdio module
